Moved package and refactoring
 - Moved package to org.jgrapht.graph.guava
 - Removed weighted variants (users can use AsWeighted view class to get same functionality)
 - Added dist package support
 - Updated bundle plugin version
 - Update readme and dependencies
