+++ jgrapht-core/pom.xml
@@ -68,7 +68,7 @@
 			<plugin>
 				<groupId>org.apache.felix</groupId>
 				<artifactId>maven-bundle-plugin</artifactId>
-				<version>3.0.1</version>
+				<version>3.5.0</version>
 				<executions>
 					<execution>
 						<phase>process-classes</phase>
@@ -80,8 +80,7 @@
 				<configuration>
 					<instructions>
 						<Automatic-Module-Name>org.jgrapht.core</Automatic-Module-Name>
-						<Bundle-Vendor>Barak Naveh, John V. Sichi and contributors (see
-							http://sourceforge.net/projects/jgrapht/)</Bundle-Vendor>
+						<Bundle-Vendor>Barak Naveh, John V. Sichi and contributors</Bundle-Vendor>
 					</instructions>
 				</configuration>
 			</plugin>
