+++ jgrapht-guava/src/main/java/org/jgrapht/graph/guava/SerializationUtils.java
@@ -15,7 +15,7 @@
  * (b) the terms of the Eclipse Public License v1.0 as published by
  * the Eclipse Foundation.
  */
-package org.jgrapht.graph;
+package org.jgrapht.graph.guava;
 
 import java.io.IOException;
 import java.io.ObjectInputStream;
