+++ jgrapht-guava/src/main/java/org/jgrapht/graph/guava/package-info.java
@@ -1,0 +1,4 @@
+/**
+ * Guava adapters. 
+ */
+package org.jgrapht.graph.guava;
