--- jgrapht-guava/src/main/java/org/jgrapht/graph/package-info.java
@@ -1,4 +1,0 @@
-/**
- * Implementations of various graphs.
- */
-package org.jgrapht.graph;
