+++ jgrapht-guava/src/test/java/org/jgrapht/graph/guava/SerializationTestUtils.java
@@ -15,7 +15,7 @@
  * (b) the terms of the Eclipse Public License v1.0 as published by
  * the Eclipse Foundation.
  */
-package org.jgrapht.graph;
+package org.jgrapht.graph.guava;
 
 import java.io.ByteArrayInputStream;
 import java.io.ByteArrayOutputStream;
