+++ jgrapht-io/pom.xml
@@ -49,7 +49,6 @@
 					</transformers>
 					<shadedArtifactAttached>true</shadedArtifactAttached>
 					<shadedClassifierName>uber</shadedClassifierName>
-					<!-- <finalName>${project.artifactId}-${project.version}-uber</finalName> -->
 				</configuration>
 			</plugin>
 			<plugin>
@@ -70,7 +69,7 @@
 			<plugin>
 				<groupId>org.apache.felix</groupId>
 				<artifactId>maven-bundle-plugin</artifactId>
-				<version>3.0.1</version>
+				<version>3.5.0</version>				
 				<executions>
 					<execution>
 						<phase>process-classes</phase>
@@ -82,8 +81,7 @@
 				<configuration>
 					<instructions>
 						<Automatic-Module-Name>org.jgrapht.io</Automatic-Module-Name>
-						<Bundle-Vendor>Barak Naveh, John V. Sichi and contributors (see
-							http://sourceforge.net/projects/jgrapht/)</Bundle-Vendor>
+						<Bundle-Vendor>Barak Naveh, John V. Sichi and contributors</Bundle-Vendor>
 					</instructions>
 				</configuration>
 			</plugin>
